ADD 레시피 추가 화면

+++ app/src/main/java/com/example/recipebook/ui/detail/RecipeDetailViewModel.kt
@@ -7,7 +7,7 @@ import java.lang.IllegalArgumentException
 
 class RecipeDetailViewModel(private val recipeDao: RecipeDao) : ViewModel() {
 
-    fun retrieveRecipe(id: Int): LiveData<RecipeWithIngredients> {
+    fun retrieveRecipe(id: Long): LiveData<RecipeWithIngredients> {
         return recipeDao.getRecipeWithIngredients(id).asLiveData()
     }
 }
